localization

+++ src/main/java/model/dao/api/GenericDao.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 
 public interface GenericDao<T> {
     void create(T entity);
-    Optional<T> findById(int id);
+    Optional<T> findById(long id);
     List<T> findAll();
     void update(T t);
-    void delete(int id);
+    void delete(long id);
 }
